requirements_develop.txt
Pillow>=9.5.0
PyYAML
Shapely==1.8.5.post1
black
colorama
flake8
flake8-html
genbadge
geojson==3.0.0
histolab==0.6.0
matplotlib
natsort
numpy>1.22,<1.24
opencv_python_headless
openslide_python
pandas
pre-commit
pydantic==1.10.4
pytest
pytest-sugar
rasterio==1.3.5.post1
requests
scikit-image
setuptools<=65.6.3
tqdm
torchvision
torch